Crate posix_space

A small crate which determines if a byte is classified as a space in the POSIX locale per POSIX.1-2017, chapter 7, Locale.

space

Define characters to be classified as white-space characters.

In the POSIX locale, exactly <space>, <form-feed>, <newline>, <carriage-return>, <tab>, and <vertical-tab> shall be included.

The function defined in this crate should have equivalent behavior to the C function isspace as defined in ctype.h.

Functions

is_space

Determine whether the given byte is in space POSIX character class.